+++ lib/ui/pages/my_library/my_libray_main_page/my_library_main_reading_note/my_library_main_reading_note_form/my_library_main_reading_note_book.dart
@@ -5,9 +5,9 @@ import 'package:flutter_blog/_core/constants/font.dart';
 import 'package:flutter_blog/_core/constants/size.dart';
 import 'package:flutter_blog/data/model/book.dart';
 
-class MyLibraryMainReadingNoteBook extends StatelessWidget {
+class MyLibraryMainReadingNotePostBook extends StatelessWidget {
   final Book book;
-  const MyLibraryMainReadingNoteBook({required this.book});
+  const MyLibraryMainReadingNotePostBook({required this.book});
 
   @override
   Widget build(BuildContext context) {
