+++ lib/ui/pages/my_library/my_libray_main_page/widgets/my_library_main_tab_bar_view.dart
@@ -4,10 +4,10 @@ import 'package:flutter_blog/data/model/board.dart';
 import 'package:flutter_blog/data/model/book.dart';
 import 'package:flutter_blog/data/model/book_reply.dart';
 import 'package:flutter_blog/data/model/user.dart';
-import 'package:flutter_blog/ui/pages/my_library/my_libray_main_page/my_library_main_reading_note/my_library_main_reading_note_form/my_library_main_reading_note_content.dart';
-import 'package:flutter_blog/ui/pages/my_library/my_libray_main_page/my_library_main_reading_note/my_library_main_reading_note_form/my_library_main_reading_note_form.dart';
-import 'package:flutter_blog/ui/pages/my_library/my_libray_main_page/widgets/my_library_main_bookcase.dart';
-import 'package:flutter_blog/ui/pages/my_library/my_libray_main_page/widgets/my_library_main_like_books.dart';
+import 'package:flutter_blog/ui/pages/my_library/my_libray_main_page/my_library_main_reading_note/my_library_main_reading_note_form/my_library_main_reading_note_reply_form.dart';
+import 'package:flutter_blog/ui/pages/my_library/my_libray_main_page/my_library_main_reading_note/my_library_main_reading_note_form/my_library_main_reading_note_post_form.dart';
+import 'package:flutter_blog/ui/pages/my_library/my_library_choice_page/my_library_main_bookcase.dart';
+import 'package:flutter_blog/ui/pages/my_library/my_library_choice_page/my_library_main_like_books.dart';
 import 'package:flutter_blog/ui/widgets/button/custom_category_button.dart';
 import 'package:flutter_blog/ui/widgets/scroll_view/custom_book_grid_view.dart';
 
@@ -159,7 +159,7 @@ class _MyLibraryMainTabBarViewState extends State<MyLibraryMainTabBarView> {
 
                               return Column(
                                 children: [
-                                  MyLibraryMainReadingNoteForm(
+                                  MyLibraryMainReadingNoteReplyForm(
                                     book: book,
                                     oneReviewComent:
                                         "${bookReplys[index].content}",
@@ -178,13 +178,13 @@ class _MyLibraryMainTabBarViewState extends State<MyLibraryMainTabBarView> {
                               final book = books[index];
 
                               return InkWell(
-                                onTap: (){
+                                onTap: () {
                                   /// TODO 대욱 : 임시주소 -> 포스트디테일 페이지로 연결
                                   Navigator.pushNamed(context, "/searchMain");
                                 },
                                 child: Column(
                                   children: [
-                                    MyLibraryMainReadingNoteContent(
+                                    MyLibraryMainReadingNotePostForm(
                                       book: book,
                                       oneReviewComent:
                                           "${bookReplys[index].content}",
